+++ src/main/java/com/sotatek/cardano/common/entity/TxOut.java
@@ -5,7 +5,7 @@ import com.sotatek.cardano.common.validation.Hash28Type;
 import com.sotatek.cardano.common.validation.Hash32Type;
 import com.sotatek.cardano.common.validation.Lovelace;
 import com.sotatek.cardano.common.validation.TxIndex;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Objects;
 import javax.persistence.Column;
@@ -78,7 +78,7 @@ public class TxOut extends BaseEntity {
   @Column(name = "value", nullable = false, precision = 20)
   @Lovelace
   @Digits(integer = 20, fraction = 0)
-  private BigDecimal value;
+  private BigInteger value;
 
   @Column(name = "token_type", nullable = false)
   private TokenType tokenType;
